backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
openai==1.52.0
pdfplumber==0.11.4
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.2

# Optional — uncomment to enable
# pinecone-client==3.2.2
# faster-whisper==1.0.3
